# fix(performance): load envinfo on demand, disable ts-node